+++ data/scripts/player/missions/cw_forcerecon.lua
@@ -78,7 +78,7 @@ function initialize(factionIndex)
         local baseReward = math.floor(75000 + heat * 100000)
 
         mission.data.reward = precomputedReward or {
-            credits = baseReward * Balancing.GetSectorRewardFactor(x, y),
+            credits = baseReward * Balancing.GetSectorRewardFactor(x, y) * ((Faction(mission.data.custom.giverIndex or 0) and Faction(mission.data.custom.giverIndex or 0):getValue("cosmic_trait_cw_mercantile") == 1) and 3 or 1),
             relations = 3000,
             paymentMessage = "Data received loud and clear. Good work out there, captain. Payment transferred."%_T
         }
@@ -195,7 +195,9 @@ function getBulletin(station)
     if heat < 0.15 then return end
 
     local baseReward = math.floor(75000 + heat * 100000)
-    local rewardCredits = baseReward * Balancing.GetSectorRewardFactor(Sector():getCoordinates())
+    local giverFaction = Faction(station.factionIndex)
+    local mult = (giverFaction and giverFaction:getValue("cosmic_trait_cw_mercantile") == 1) and 3 or 1
+    local rewardCredits = baseReward * Balancing.GetSectorRewardFactor(Sector():getCoordinates()) * mult
     local rewardStruct = {
         credits = rewardCredits,
         relations = 3000,