+++ data/scripts/player/missions/cw_propaganda_broadcast.lua
@@ -61,7 +61,7 @@ function initialize(factionIndex)
         local baseReward = math.floor(100000 + heat * 150000)
 
         mission.data.reward = precomputedReward or {
-            credits = baseReward * Balancing.GetSectorRewardFactor(x, y),
+            credits = baseReward * Balancing.GetSectorRewardFactor(x, y) * ((Faction(mission.data.custom.giverIndex or 0) and Faction(mission.data.custom.giverIndex or 0):getValue("cosmic_trait_cw_mercantile") == 1) and 3 or 1),
             relations = 10000,
             paymentMessage = "Contract fulfilled. Payment transferred."%_T
         }
@@ -132,7 +132,9 @@ function getBulletin(station)
     if heat < 0.25 then return end
 
     local baseReward = math.floor(100000 + heat * 150000)
-    local rewardCredits = baseReward * Balancing.GetSectorRewardFactor(Sector():getCoordinates())
+    local giverFaction = Faction(station.factionIndex)
+    local mult = (giverFaction and giverFaction:getValue("cosmic_trait_cw_mercantile") == 1) and 3 or 1
+    local rewardCredits = baseReward * Balancing.GetSectorRewardFactor(Sector():getCoordinates()) * mult
     local rewardStruct = {
         credits = rewardCredits,
         relations = 10000,